private contact routes

+++ controllers/contactController.js
@@ -4,16 +4,16 @@ const mongoose = require('mongoose');
 
 //@desc Get All Contact
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getAllContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id : req.user.id});
   res.status(200).json(contacts);
 });
 
 //@desc Create New Contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   // console.log("The request body : ", req.body);
   const { name, email, phone } = req.body;
@@ -25,13 +25,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 //@desc Get Single Contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   // Validasi ID sebelum melakukan query
    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
@@ -49,7 +50,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update Single Contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
     res.status(400);
@@ -62,6 +63,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if(contacts.user_id.toString() !== req.user.id){
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -71,7 +77,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Single Contact
 //@route Delete /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
     res.status(400);
@@ -82,6 +88,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if(!contacts){
     res.status(404);
     throw new Error("Contact not found");
+  }
+  
+  if(contacts.user_id.toString() !== req.user.id){
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await Contact.findByIdAndDelete(req.params.id);
